Introduction of the device driver deprecation data support (D4)
This patch introduces support for our new
device_driver_depreaction_data.json file that will contain all
information about support, deprecation and their removal of devices and
drivers.
The data will be either fetched from Insights or read from
/etc/leapp/files if present there.

In place are checks for unsupported:
- Drivers
- CPUs
- PCI Devices

In future more things might be added to the data file, however we are
currently filtering on entries for cpu and pci because we do not have
any handling for it. If this data will show up it will not affect the
code and only once we added support we can handle those too.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>