el8toel9: Add actor that ensures ifcfg files work in el9
RHEL 9 is not going to ship network-scripts. While NetworkManager
generally supports ifcfg files just fine, there are some corner cases
that are either unsupportable or explicitly configured not to work with
NetworkManager.

This adds an actor that checks for those cases and inhibits upgrade if
they are encountered.

Also, in case someone has network-scripts installed, but not
NetworkManager, make sure the proper packages needed to support the
existing configuration will end up being installed.

https://bugzilla.redhat.com/show_bug.cgi?id=2052343