Introduction of DNFWorkarounds
Previously we have had only one workaround to handle before we were
running the DNF transaction. That was implemented by executing the
'handleyumconfig' script. This was applicable for el7 => el8 upgrades.
Now we are currently having a WIP PR for handling satellite upgrades,
and the need for such a workaround has emerged there as well. The
problem is just, that it cannot be properly conditionalized the way how
it is implemented at the moment without importing too much irrelevant
things to the dnfplugin part of the workflow.
To solve this, this patch will introduce now the DNFWorkaround model,
that will allow actors to register a workaround script to be registered
when the right conditions are met and all of the workarounds are
processed as received before dnf is called. (Just like it was the case
for the handleyumconfig script before).

The advantage of this implementation is that it will now make this whole
approach more modular and can be used by 3rd parties and customers a
like to workaround their problems by just creating an actor that
registers their own script.

As part of this patch, the changes that are necessary for
handleyumconfig have been already applied.

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>