Refactor the pes_events_scanner actor
The pes_events_scanner actor is a critical part of the upgrade process,
however, its implementation degraded over time due to requirements
changing and bugs being discovered to a point that it is hard to debug
and to extend this component. This patch rewrites the old
implementation using implicit invariants imbued in the old
implementation, while trying to make them more explicit. The actor no
longer tracks what particular operation should be performed with a
package (install/keep/remove) in a manner making sure that the package
is only in one of the install/keep/remove buckets. Instead, the actor
starts with a set of installed packages and gradually changes this set -
removes/adds packages - according to the semantics of events. This shift
improves the readability of the code, making it easier to track the
effect of an event to a set of packages. The refactor also splits the
target package set computation away from event parsing, further
improving readability.