VDO RHEL8 to RHEL9 upgrade (#821)
Add scan & checks for VDO devices (+ bump leapp-framework 2.2+) (#821)

VDO devices must be converted to be managed by LVM prior the upgrade
otherwise they will not be accessible after the upgrade to rhel 9
(affects only IPU 8 -> 9). In case any VDO non-converted devices
are discovered OR we cannot check whether any such devices are present
and there is a possibility they are present, the upgrade is inhibit.
The second case could happen when lvm is installed but VDO package
not - which is required to be able to scan VDO devices. In such a case,
user will have possibility to decide what should happen answering a user
question (or they could install the vdo rpm to make the scan possible).

The vdo package is installed by default - but not in case of the minimal
installation so it's not expected to see this question so frequently as in case
of PAM during IPU 7 -> 8.

Information about VDO devices are stored in the `VdoConversionInfo` msg.

Bumped requirements to leapp-framework 2.2+ as Dialogs are broken
on Python3 for older versions.