el8toel9: Add actor that scans for deprecated network settings
RHEL 9 is going to deprecate certain network configurations. In
particular, it will ship without support for WEP networks.

Let's make sure the users are not caught by surprise that their
networks no longer connect.

Pick a generic name for the actor, as there are possibly more deprecations
coming in future.

https://bugzilla.redhat.com/show_bug.cgi?id=2056925