repomapping: add heuristics for rhui when pesid repo entries share repoids
Repositories mapping relies on the fact that repository ID precisely
identifies a repository, and therefore, there cannot be more than one
repository with the same repoid. It turns out that this might not be the
case (e.g. on some clouds) and that further logic needs to be
implemented in order to provide correct repositories on for the target
system. This patch implements heuristics for could systems that will
prefer the repositories that are native to the cloud when pesid entries
sharing the repoid were found.