[IPU 8->9] Detect RPMs with RSA/SHA1 signatures & crypto policies
tl;dr;
 * get the currennt crypto policies
 * inhibit upgrade when rpms with RSA/SHA1 signature are installed
   and used crypto policies will not allow it on the upgrade system
   * the only allowed crypto policy in such a case is LEGACY or
     DEFAULT:SHA1
 * set the crypto policy in the target userspace container if
   it is not DEFAULT

The SHA-1 hash algorithm is considered as unsafe and it is not longer
allowed to be used on RHEL 9 system using default crypto policies.
One of side effects is problem with installed packages with RSA/SHA1
signatures, as these packages could cause fail of the in-place upgrade
because such packages cannot be read, nor removed by dnf/rpm so the
tests of the transaction already fails in case we want to remove
(include replace) such packages. Additionally, even when we do not
touch such packages during the IPU, the final system will be still
malfunction as rpm will not be able to handle it by default.

Exception are packages signed by RSA/SHA1 keys that are not imported
in the system, so such signatures are not checked and are ignored
like in the case no signature is present at all.

To prevent unwanted situations regarding the installed packages
(packages that will be installed during the upgrade will be handled
in separate commit):
 * inhibit the upgrade when an RSA/SHA1 rpm is installed and
   current crypto policy is not LEGACY or DEFAULT:SHA1
   * NOTE: SHA1 subpolicy have to be created by user manually on the
           the system (and correctly!!) to make it working,
           so rather do not suggest it in remediation instructions
 * just report presence of such packages with HIGH risk so users
   are informed about the changed policy on RHEL 9

The current crypto policy is obtained from
  /etc/crypto-policies/state/current

In case the policy is not "DEFAULT", set the crypto policy inside
the targetuserspace container. However, in case of custom policies
it can be broken (policies have to be copied/installed into the
container as well).
This is currently expected limitation. It is expected to improve it
in future.

Co-authored-by: Toshio Kuratomi <a.badger@gmail.com>
Co-authored-by: Michal Reznik (mreznik) <mreznik@redhat.com>
Co-authored-by: Alexander Sosedkin <monk@unboiled.info>