call Satellite installer with --disable-system-checks if possible
The installer has a set of checks to verify whether the current system
is suitable for running Satellite. The administrator of the system can
choose to ignore those checks with `--disable-system-checks`.

As the installer invocation inside LEAPP is non-interactive, we should
err on the side of not running checks, so that the upgrade doesn't abort
in the case where the administrator has chosen to ignore the warnings.

This is in line with other non-interactive invocations of the installer
that other tools (like foreman-maintain) do.

The "if katello installer" logic is needed, as the checks and the cli
parameter is only present in Katello installations, not plain Foreman.