Fix set of supported OS versions for IPU 8 -> 9
Previously we have updated the upgrade_paths.json file to create
a mapping 8.7 -> 9.0. However, we have not enabled 8.7 for in-place
upgrades. Also, RHEL for SAP HANA has not been enabled for
IPU 8 -> 9.

Enable 8.7 for rhel and 8.6 for saphana for IPU 8 -> 9 via
SUPPORT_VERSIONS in the leapp.libraries.common.config.version
library.