Enable gpgcheck during IPU (+ add --nogpgcheck CLI option)
Previously the gpgcheck=0 has been enforced during the IPU as we have
not have any reasonable solution doing upgrade with allowed gpgcheck.
Previously such upgrades automatically imported any gpgkeys without
any possible check whether the actual keys are valid or not, which
could lead to the automatical import of compromised/spurious gpg keys
and user would not know about that. So using the original solution,
user was asked for the import of new keys when installing additional
content from new repositories (if keys have been different from those
used in the original system).

To do the upgrade in the original way (without gpgcheck), execute
leapp with the `--nogpgcheck` option, or specify `LEAPP_NOGPGCHECK=1`
(envar). In such a case, all actions described below are skipped.

The current solution enables the GPG check by default but also could
require additional actions from user to be able to upgrade. The goal
is to ensure that no additional GPG keys are imported by DNF
automatically during any action (partially resolved, read bellow).
To be able to achive this, we are importing gpg keys automatically
from a *trusted directory* before any dnf transaction is executed, so:
 a) into the rpmdb of the target userspace container, before the
    actual installation of the target userspace container
 b) into overlayed rpmdb when calculating/testing the upgrade
    transaction
 c) into the system rpmdb right before the execution of the DNF
    upgrade transaction

The case a) is handled directly in the target_userspace_creator actor.
The other cases are handled via DNFWorkaround msg, using the
`importrpmgpgkeys` tool script.

The *trusted directory* is in this case located under
`files/rpm-gpg/`, where the directory name is major release of the
target system in case of production releases, in other cases it has
the *beta* suffix. So e.g.:
    files/rpm-gpg/8
    files/rpm-gpg/8beta
That's because production and beta repositories have different gpg
keys and it is not wanted to mix production and beta keys. Beta
repositories are used only when the target production type is beta:
    LEAPP_DEVEL_TARGET_PRODUCT_TYPE=beta

Introducing the missinggpgkeysinhibitor actor that checks gpg keys
based on `gpgkey` specified in repofiles per each used target
repository which does not explicitly specify `gpgcheck=0`.
Such a key is compared with the currently installed gpg keys in the
host rpmdb and keys inside the *trusted directory*. If the key
is not stored in any of those places, the upgrade is inhibited with
the corresponding report. User can resolve the problem installing the
missing gpg keys or storing them to the trusted directory.

Currently supported protocols for the gpg keys are
  file:///
  http://
  https://
If a key cannot be obtained (including use of an unsupported protocol,
e.g. ftp://) the actor prompt a log, but does not generate a report
about that (so the upgrade can continue, which could later lead into
a failure during the download of packages - one of TODOs).

This is not the final commit for this feature and additional work
is expected before the new release is introduced. Regarding that,
see the code for new TODO / FIXME notes that are put into the code.

Summary of some TODOs planned to address in followup PR:
  - add checks that DNF does not import additional GPG keys during
    any action
  - report GPG keys that could not be checked, informing user about
    possible consequences - the report should not inhibit the upgrade
  - possibly introduce fallback for getting file:///... gpg keys
    as currently they are obtained from the target userspace container
    but if not present, the host system should be possibly checked:
    - Note that if the file has been created manually (custom repo file)
      most likely the gpgkey will be stored only on the host system
      - and in such a case the file would need to be copied from the
        host system into the container.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>