StorageInfo: Extend LsblkEntry model by bsize and kname
The current data from lsblk contains `size` in human readable format
which is not ideal format for the work in python. Introducing `bsize`
field that contains the size in bytes.

Also for some usecases we have discovered that it could be good
to know the kernel name of specific device. So introducing `kname`
field providing this value.