Add Makefile target for building locally in containers
Builds are done in containers where dependencies can be met, for example
`python-devel` (python2) on el7, which might not be available on host
machine.

For el7 builds CentOS 7 container is used, because `rpm-build` package
is not available on ubi7.

To build RPMs in a container use `make build_container`. The
`BUILD_CONTAINER` variable must be set to either `el7` or `el8`.

Also added .gitignore entries for subfolders in `packaging/` as
packages and intermediate build files shouldn`t be in the repo.