BZ#2111835 - move PostgreSQL data content individually
Some users prefer to have the data directory on an own mountpoint (vs
the whole lib/pgsql directory). In such setups we cannot just move data
to the new location, as we can't move the mointpoint.
Instead, lets move the contents of the data directory individually,
which should work fine in this scenario.

Also wrap all code that can fail with try/except, so that we can report
errors but don't fail the overall execution of the upgrade.