Fix swap page size during el8toel9 upgrade on aarch64
The kernel in RHEL8 uses 64k page size, but this is changed in RHEL9 to
4k. This breaks SWAP and it must be reinitialized.

The reinitialization is done in initramfs before first boot into the new
system as on some systems SWAP is essential to boot successfully.

The `swapon` command used for reinitialization internally uses
`mkswap` and both of these commands are unavailable in the dracut shell.
As a workaround we can use `$NEWROOT/usr/sbin/swapon` and
`$NEWROOT/usr/sbin/mkswap`. For `swapon` to find `mkswap` the `PATH` must be
temporarily adjusted.

Co-authored-by: Petr Stodůlka <xstodu05@gmail.com>