Mass fix of imports in actors to make linters happy (disable isort)
As we are doing already a mass refactoring across actors, let's deal
with this problem as well.

This commit fixes majority of problems reported by isort. However
in some cases isort is requiring a change even when imports are
correct regarding our policies (any python standards as well).
The required changes lead into not so nice set of imports
(e.g. split grid import into several imports when a comment
is present, etc).

As we consider this commit useful to merge, let's for now temporarily
disable isort (re-enable it in the following commit). Let's keep
the better solution for future.