Provide common information about systemd
Introduced new actors providing information about systemd services,
preset files, and broken symlinks. Both for the source and the
target system. All related actors are under
    repos/system_upgrade/common/actors/systemd/
directory. Also it's introduced the systemd shared library
providing basic functionality for the gathering of systemd data.
Currently all functions from the library that are not called by
actors directly are marked as private. Note that provided functions
could raise exceptions. The handling of exceptions is expected to be
done inside actors. See docstrings for the set of possible exceptions.

In case of the source system, all data is valid during the upgrade
and in case the required data cannot be obtained, the upgrade
is interrupted.

However in case of data about the target system we speak about
a snapshot, how the system looks like in the moment after the upgrade
rpm transaction (data is collected during the Application phase).
The data can be used for the basic overview of the system configuration
after the upgrade transaction and also does not have to be necessary
produced! In case of an error a particular *Target* msg is not
produced, to make clear we could not collect correctly the required
data. But the upgrade is not interrupted in this phase and the errors
are logged only.

Systemd symlinks (under /etc/systemd/system/) prior the upgrade are
reported during the preupgrade, so administrator could fix them
prior the upgrade.

It's expected to create post-upgrade reports in future, but currently
skipping this topic until the post-upgrade reports are defined by
the leapp framework.

Introduced models (messages):
 * SystemdBrokenSymlinksSource
 * SystemdServicesInfoSource
 * SystemdServicesPresetInfoSource

 * SystemdBrokenSymlinksTarget
 * SystemdServicesInfoTarget
 * SystemdServicesPresetInfoTarget