Add an actor that enables device_cio_free.service on s390x
After an IPU the device_cio_free.service systemd service (exclusive to
s390x) is disabled even though the vendor preset is set to disable. The
new actor instruct to enable the service during the IPU.

The service is used to enable devices not explicitly enabled on kernel
command line.

Jira ref.: OAMG-6302