Add the scanzfcp actor handling the IPU with ZFCP (s390x)
When having systems configured with ZFCP instead of DASD, the disks
are not seen while rebooting because `/etc/zfcp.conf` is missing
in the initramfs.

When the file exists, it's copied inside the userspace container
and installed in the upgrade initramfs, producing
   TargetUserSpaceUpgradeTasks, UpgradeInitramfsTasks
messages.

pstodulk: updated unit-tests in the scanzfcp and scandasd actors

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2140563

Signed-off-by: Renaud Métrich <rmetrich@redhat.com>