Fix broken or incorrect systemd symlinks
Introduce repairsystemdsymlinks actor.
During the in-place upgrade process, it usually happens that some
symlinks become incorrect - symlinks are broken, or they are defined
in a wrong directory (e.g. when they are supposed to be defined in a
different systemd target). This has various reasons, but usually it's
caused by missing rpm scriptlets in particular rpms.

This change corrects only systemd symlinks are (newly) broken during
the in-place upgrade. Symlinks that have been already broken before
the in-place upgrade are ignored.

Symlinks are handled in the following fashion, if the symlink points to:
    - a removed unit, such a symlink is deleted
    - a unit whose installation has been changed (e.g. changed WantedBy),
      such symlinks are fixed (re-enabled using systemctl)

JIRA:
  OAMG-5342
  OAMG-5344
  OAMG-6519 (possibly related)
  OAMG-7755

Bugzillas:
  https://bugzilla.redhat.com/show_bug.cgi?id=1988457
  https://bugzilla.redhat.com/show_bug.cgi?id=1988449
  https://bugzilla.redhat.com/show_bug.cgi?id=2055117 (possibly fixed)