Use "grub" library to find the GRUB device
We need to find the device with GRUB in order to update grub core
(on BIOS systems). Previously, we used "grubdevname" actor to get the
info and send a message but this solution was error prone due to device
renaming.

This fixes only the "checkgrubcore" actor as the "updategrubcore" actor
already uses the library.

Co-authored-by: Michal Hecko <mhecko@redhat.com>