Fix certificates blacklisting for rhel8->rhel9 upgrade
If you blacklist a certificate on rhel8 and then upgrade to rhel9, the
certificate does not remain blacklisted.
On rhel9 we renamed blacklist to blocklist, so after the upgrade we have
both blacklist and blocklist folders but the certificates are not moved to
the blocklisted folder.

The actor migrateblacklist.py was created to solve this issue, but it did not
create messages and reports.
This checkin adds scan and check versions which generate the appropriate
messages and reports.

migrateblacklist.py has not been updated to use these messages yet.