Automatically register systems with RHSM into Red Hat Insights
If system uses rhsm it is automatically registered into Red Hat Insights
at the end of the upgrade process.
For the registration the `insights-client` package is requiered and will
be installed during the upgrade if missing.

This can be skipped by setting the `LEAPP_NO_INSIGHTS_REGISTER=1`
environment variable or `--no-insights-register` CLI option.

A report is generated informing about the registration and eventual
`insights-client` package installation.

Jira ref.: OAMG-7872