Packit: specify srpm_build_deps to allow builds with Packit again
The option / list is mandatory now for Packit, however we can keep
it empty for now as it looks like as we build the srpm via the
Makefile still. The same is used now in the leapp project.

It's expected we update the list in upcoming months most likely, when
we integrate this project better with the current Packit best practices.