scaninstalledtargetkernelversion: Detect kernel-core package on RHEL 8+
On RHEL 8+ the `kernel` package was split into multiple
packages including `kernel-core`. The `kernel` package is just a
placeholder there and might not be installed.

The `kernel-rt` package was split similarly.

Let's scan for `kernel-core` and (`kernel-rt-core` on RHEL Real Time)
instead which should always be present.

Jira ref.: OAMG-8383