Detect RoCE on IBM Z (s390x) machines
When upgrading RHEL 8 to RHEL 9 on IBM Z systems,
it's important to have correctly configured RoCE
if used, otherwise the network will be down after
the upgrade as the default configuration is not
persistent between RHEL 8 and RHEL 9.

Currently it's possible to configure RoCE in a persistent
way for both RHEL systems since RHEL 8.7 by following
kernel cmdline argument:
    net.naming-scheme=rhel-8.7
RHEL 8.7 and newer has persistent RoCE NICs if booted
with the parameter.

For the detection of active RoCE NICs is used the nmcli tool
(considerring that RoCE needs to be configured via NetworkManager
on RHEL 8+). Detected active RoCE NICs are represented by
the RoceDetected msg.

The upgrade is inhibited if any active RoCE NIC is discovered and:
* source os is RHEL 8.6 or older, or
* the system is booted without `net.naming-scheme=rhel-8.7` on kernel
  cmdline

Co-authored-by: Petr Stodulk <pstodulk@redhat.com>