fix migration of sendmail configs after the upgrade
The migrate_sendmail actor would cause a traceback if a file marked for
migration wasn't found, which would be caused for example by removing
the sendmail package during the upgrade. The actor now checks for
existence of files that need to be migrated and also informs about files
that could not be migrated.